pubmed-company-papers/pyproject.toml
[tool.poetry]
name = "pubmed-company-papers"
version = "0.1.0"
description = "A tool to fetch research papers from PubMed with authors affiliated with pharmaceutical or biotech companies"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "pubmed_company_papers"}]

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.81"
pandas = "^2.0.0"
tqdm = "^4.65.0"
typing-extensions = "^4.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[tool.poetry.scripts]
get-papers-list = "pubmed_company_papers.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true